requirements.txt
Pillow==10.3.0
playsound==1.3.0
pyzbar==0.1.9
Requests==2.32.3
urllib3==1.26.18
win11toast==0.32
